Closes #414: Move IWebViewLifecycleF to web package.
imo, the fragment/ package doesn't serve any value because it doesn't
group related code. We could conceivable merge the web/ and browser/
packages together but I'll reserve this for the end of our refactor.
