+++ app/src/main/java/org/mozilla/focus/fragment/InfoFragment.java
@@ -16,6 +16,7 @@ import org.mozilla.focus.R;
 import org.mozilla.focus.session.NullSession;
 import org.mozilla.focus.session.Session;
 import org.mozilla.focus.web.IWebView;
+import org.mozilla.focus.web.IWebViewLifecycleFragment;
 
 public class InfoFragment extends IWebViewLifecycleFragment {
     private ProgressBar progressView;
